changed weight counter to float for proper atomic support

+++ mitsuba/src/integrators/path/guided_path.cpp
@@ -912,7 +912,7 @@ struct DTreeWrapper {
 public:
     DTreeWrapper() : m_rejPdfPair(1.f, 1.f){
         current_samples = 0;
-        weighted_previous_samples.compare_exchange_weak(weighted_previous_samples.load, 0);
+        weighted_previous_samples.compare_exchange_weak(weighted_previous_samples.load, 0.f);
         req_augmented_samples = 0;
         B = 0.f;
         min_nzradiance = std::numeric_limits<float>::max();
@@ -1127,7 +1127,7 @@ private:
 
     std::uint64_t current_samples;
     std::uint64_t req_augmented_samples;
-    std::atomic<double> weighted_previous_samples;
+    std::atomic<float> weighted_previous_samples;
     float B;
 
     std::pair<Float, Float> m_rejPdfPair;
